+++ src/main/java/com/calculator/MainVerticle.java
@@ -1,12 +1,14 @@
 package com.calculator;
 
 import io.vertx.core.AbstractVerticle;
+import io.vertx.core.json.JsonObject;
 import io.vertx.ext.web.Router;
 import io.vertx.ext.web.RoutingContext;
 import io.vertx.ext.web.openapi.RouterBuilder;
 import io.vertx.ext.web.validation.ParameterProcessorException;
 import io.vertx.ext.web.validation.RequestParameters;
 import io.vertx.ext.web.validation.ValidationHandler;
+import io.vertx.jdbcclient.JDBCPool;
 
 public class MainVerticle extends AbstractVerticle {
 
@@ -35,10 +37,10 @@ public class MainVerticle extends AbstractVerticle {
 
     RequestParameters requestParameters = routingContext.get(ValidationHandler.REQUEST_CONTEXT_KEY);
 
-    Integer number1 = requestParameters.queryParameter("number1").getInteger();
-    Integer number2 = requestParameters.queryParameter ("number2").getInteger();
+    Double number1 = requestParameters.queryParameter("number1").getDouble();
+    Double number2 = requestParameters.queryParameter ("number2").getDouble();
 
-    int resultado = Calculator.sum(number1, number2);
+    double resultado = Calculator.sum(number1, number2);
     ResultDTO resultadoDTO = new ResultDTO(resultado);
     routingContext.json(resultadoDTO);
   }
